// pathlabmasterfrontend/src/app/super-admin/parameter/parameter-types.ts
export type ApiParameter = {
  parameterId?: number | string;
  parameterName?: string;
  code?: string;
  value?: string;
  sequence?: number | string;
  dataType?: string;
  unit?: string;
  criteria?: string;
  defaultVlue?: string;
  formula?: string;
  upperRange?: number | string | null;
  lowerRange?: number | string | null;
  extrimUpperRange?: number | string | null;
  extrimLowerRange?: number | string | null;
  lowerAgeRange?: number | string | null;
  upperAgeRange?: number | string | null;
  method?: string;
  context?: string;
  isHideLable?: boolean;
  isHideLableOnRemport?: boolean;
  isLocalDictonery?: boolean;
  isWrapper?: boolean;
  isCalculative?: boolean;
  isImageResize?: boolean;
  isBold?: boolean;
  createdBy?: number | string;
  updatedBy?: number | string;
  createdAt?: string;
  updatedAt?: string;
};